feat: DH-16956: Move fontbootstrap from app-utils to components so it can be consumed in DHE (#2545)
To fix DH-16956 and also resolve weird Pop OS linux font issues, we
should switch enterprise to use the fontsource fonts. Since app-utils is
not used in enterprise and can't be imported there, moving fontbootstrap
to components under theme seemed like the best solution.

Will need a change in DHE to adopt font-bootstrap once published, and a
heads up to enterprise QA that there may be minor e2e test differences
similar to what we had in core in #1944