feat: Virtualize column organization lists (#2568)
Fixes #1650

Further improvement for DH-18960 that wasn't part of the ticket, but
will be appreciated. The original ticket was big enough I didn't want to
add another feature to that PR.

Fixes the issue when dragging lots of columns the drag overlay grew. It
is now properly truncating. Also removed the drag handle icon from the
drag overlay just to give a bit more room when dragging groups (and I
think the handle doesn't make that much sense to show when you're
already dragging)

Also fixed the drop animation when dragging from/to a group. Same issue
as the general `movedColumns` fixed in the first PR.