fix: DH-20243: improved notebook settings merging (#2593)
It is expected that `getSettings` substitutes custom settings on top of
default settings when the settings are pulled.
The problem is this substitution was only happening properly if top
level settings were being replaced.
This meant that the nested linter settings were not getting properly
replaced for `notebookSettings` because `notebookSettings` is never
itself `undefined`, only the keys `config` and `isEnabled` are.